+++ src/driver_EIC.h
@@ -1,24 +1,24 @@
 #pragma once
 
-#include <stdbool.h>
+#include <stddef.h>
 
 typedef struct EIC_Cfg_ {
-  int ch;
-  int sense;
-  int pin;
-  void (*cb)();
+  size_t  ch;    /* Channel index */
+  uint8_t sense; /* Sense direction */
+  size_t  pin;   /* Pin to attach */
+  void (*cb)();  /* Callback on interrupt */
 } EIC_Cfg_t;
 
 /*! @brief Set the callback function for an EIC channel
  *  @param [in] ch : EIC channel
  *  @param [in] cb : callback function pointer
  */
-void eicCallbackSet(const int ch, void (*cb)());
+void eicCallbackSet(const size_t ch, void (*cb)());
 
 /*! @brief Disable EIC level line
  *  @param [in] ch : EIC channel to disable
  */
-void eicChannelDisable(const int ch);
+void eicChannelDisable(const size_t ch);
 
 /*! @brief Enable EIC level line with optional callback
  *  @param [in] eiccfg : EIC channel configuration
@@ -37,7 +37,7 @@ void eicSetup(void);
 /*! @brief Setup a pin for EIC function
  *  @param [in] pin : (logical) pin to configure
  */
-void eicPinSetup(const int pin);
+void eicPinSetup(const size_t pin);
 
 /*! @brief Setup the External Interrupt Controller for the pulse controller  */
 void eicSetupPulse(void);
